+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Diskover Change Log
 
-## [1.5.0-rc10] = 2018-06-15
+## [1.5.0-rc10] = 2018-06-16
 ### notice
 - Amazon S3 inventory support is beta, requires diskover-web >= v1.5.0-rc9
 - --s3 requires index named diskover-s3-indexname
@@ -20,7 +20,9 @@
 - moved autotag code after plugin code when setting file/directory doc meta data fields
 - main diskover.py dispatcher does not exit when there are dir size jobs still in Redis diskover_crawl queue
 - progress bar appearance
+- set default for shards/replicas to 1/0 in diskover.cfg.sample (most users are just using single es node)
 ### fixed
+- forcemerge/optimize at end of crawl causing traceback exception
 - bugs with autotagging
 - crawlbot continuous scanner (-B) strack trace error (logger)